+++ src/test/java/school/redrover/runner/TestUtils.java
@@ -1,12 +1,11 @@
-package school.redrover;
+package school.redrover.runner;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 public final class TestUtils {
-    private static final By BY_USER_ID = By.xpath("//a[contains(@href, 'user')]");
 
     public static String getUserID(WebDriver driver) {
-        return driver.findElement(BY_USER_ID).getText();
+        return driver.findElement(By.xpath("//a[contains(@href, 'user')]")).getText();
     }
 }
